backend/requirements.txt
openai
google-generativeai
python-dotenv
google-genai
flask
flask-cors
PyPDF2
gunicorn
waitress
spacy
scikit-learn
joblib
pandas